Feature #5 (#6)
Implementing work for autopop flag working for the go_to_mark function
call. Includes some clean up as well of calling the windower to close
the window instead of manually performing the action.

Fixed a bug where moving a mark up the list to the top of the list and
them removing it with autopop would cause an index out of bounds by
subtracting 1 from the currentMarkHandle leading to a value of 0.

Co-authored-by: michaelplatt07 <michael@lugia.lugia>